[FLINK-23192][table] Move DataGen connector to a proper package
