+++ flink-table/flink-table-api-java-bridge/src/main/java/org/apache/flink/connector/datagen/table/datagen/DataGeneratorContainer.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.factories.datagen;
+package org.apache.flink.connector.datagen.table.datagen;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.configuration.ConfigOption;
