+++ flink-table/flink-table-api-java-bridge/src/main/java/org/apache/flink/connector/datagen/table/datagen/RandomGeneratorVisitor.java
@@ -16,22 +16,22 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.factories.datagen;
+package org.apache.flink.connector.datagen.table.datagen;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.configuration.ConfigOption;
 import org.apache.flink.configuration.ConfigOptions;
 import org.apache.flink.configuration.ReadableConfig;
+import org.apache.flink.connector.datagen.table.DataGenConnectorOptionsUtil;
+import org.apache.flink.connector.datagen.table.datagen.types.DataGeneratorMapper;
+import org.apache.flink.connector.datagen.table.datagen.types.DecimalDataRandomGenerator;
+import org.apache.flink.connector.datagen.table.datagen.types.RowDataGenerator;
 import org.apache.flink.streaming.api.functions.source.datagen.DataGenerator;
 import org.apache.flink.streaming.api.functions.source.datagen.RandomGenerator;
 import org.apache.flink.table.api.ValidationException;
 import org.apache.flink.table.data.GenericArrayData;
 import org.apache.flink.table.data.GenericMapData;
 import org.apache.flink.table.data.StringData;
-import org.apache.flink.table.factories.DataGenConnectorOptionsUtil;
-import org.apache.flink.table.factories.datagen.types.DataGeneratorMapper;
-import org.apache.flink.table.factories.datagen.types.DecimalDataRandomGenerator;
-import org.apache.flink.table.factories.datagen.types.RowDataGenerator;
 import org.apache.flink.table.types.logical.ArrayType;
 import org.apache.flink.table.types.logical.BigIntType;
 import org.apache.flink.table.types.logical.BooleanType;
