+++ flink-table/flink-table-api-java-bridge/src/main/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -13,6 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.flink.table.factories.DataGenTableSourceFactory
+org.apache.flink.connector.datagen.table.DataGenTableSourceFactory
 org.apache.flink.table.factories.BlackHoleTableSinkFactory
 org.apache.flink.table.factories.PrintTableSinkFactory
