+++ flink-table/flink-table-api-java-bridge/src/test/java/org/apache/flink/connector/datagen/table/datagen/types/DecimalDataRandomGeneratorTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.factories.datagen.types;
+package org.apache.flink.connector.datagen.table.datagen.types;
 
 import org.apache.flink.table.data.DecimalData;
